+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Annotations
 - Conversion of `array|object` types by `BuiltinTypeSafeConverter`.
+- Deprecated configuration option `entity_response.default_content_type`.
 
 ## [1.4.2] - 2022-01-03